Split the bulk-action tests into their own suite
The three cases added in the previous commit put RuleSelectionDialogTests past
the 300-line type-body limit. They get their own file rather than a lint
exclusion, the same way the Direct Instantiation suites were split.
